Fix tracing status codes on error.
By allowing the error handler middleware to pass
the error up, the opentracing middleware can access
it and add it to the trace.

By moving the opentracing middleware after the
error handling middleare, the response will have
been written and the trace status code will be
correct.

NOTE: I'm pretty convinced at this point that we should
remove the error handling middleware and provide helper
methods instead, to be used in the endpoint handlers.
As an added benefit, there will be no more need to return an error
so we can conform to the http.Handler interface.