Ensure JSONDecoder reads all the way to EOF
JSONDecoder would not always call .Read() enough times to get an EOF.
When this happens, the connection is not returned to the connection
pool. Instead, it's closed with data still in the receive buffer, which
results in the kernel sending a TCP RST.

This is happening nearly every time in r101-experiments, which is really
bad for performance.

Further reading: https://ahmet.im/blog/golang-json-decoder-pitfalls/

The author of that article argues for using json.Unmarshal, but I
believe the concerns can be addressed by using json.Decoder.More() while
retaining the RAM advantages of a streaming parser.