Minor improvements to the httpx/errors testing and implementation
Switched the test to a whitebox (we need to test that some private functions). Added some tests around the genStacktrace funcs on boundries for fetching stack trace frames.

Swapped the order for the Recover function switch statement so it would hit the *Error case.

Switch httpx/middleware/timeout ErrHandlerTimeout to be a pointer to the error. This enabled proper interface comparisons in switch statements and casting.

Minor nit in request.go around copying a string array, which should use `copy` instead of a for loop.