Swap stvp rollbar library for official go rollbar lib
The rollbar library we've been using previously has been archived in
favour of an official rollbar go library that rollbar maintains. This
updates our usage of the rollbar library to use the active library. This
also involved minor fixes to update library calls to match the api of
the new library.