Google Cloud Profiler: Never error on startup
We've had multiple issues caused by misconfigurations in the google
cloud profiling crashing an app at startup. Because of the way this
package is used with urfave/cli, it's not realistically feasible for the
application developer to handler errors in the app itself.

These changes log any errors that occur, but ensure that no error will
be returned and the application will be able to start up without
failing.